master/requirements.txt
futures
grpcio==1.41.0
grpcio-tools==1.41.0
protobuf==3.18.1
six==1.16.0
